[SQL] Moved aggregate implementations to separate files
