+++ sql-to-dbsp-compiler/SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/sql/simple/IncrementalRegressionTests.java
@@ -56,7 +56,7 @@ public class IncrementalRegressionTests extends SqlIoTest {
                 create table P(a varchar, ts TIMESTAMP);
                 create view win AS
                 SELECT ROW_NUMBER() OVER (PARTITION BY a, HOUR(ts) ORDER BY ts DESC)
-                FROM P;""", "WINDOW aggregate with ROWS/ROW_NUMBER not yet implemented");
+                FROM P;""", "Aggregate function ROW_NUMBER not yet implemented");
     }
 
     @Test
